First pass implementation of rasterset using rioxarray
Use rioxarray instead of rasterio to read raster files.  This was a
fairly straightforward translation as my recent changes prepared the
way.

For now keeping single band interface but it should be possible to
operate on multiple bands at once so we can project in time in one
operation.

Not yet using dask for parallelism.