Initial implementation of rasterset evaluation using dask
Added a new method (build) which returns a dask computation graph.

The dask graph can be executed with load(), compute() or persist().  Or
we can build up on it, for example, to generate a zarr of netcdf file.

For now the old API (eval) still works but I haven't tried write() and
swrite() because I haven't setup tests that use high-resolution rasters.

There are a couple of small issues to iron out, e.g. the final result is
current 3-D instead of 2-D.