Don't implement raster based on rioxarray
The dask scheduler seems rather primitive and doesn't generate a "nice"
schedule when the IO happens in one task and the computation in another.

So I plan to merge the IO with the computation into a single node in the
scheduling graph.  This requires getting rid of rioxarray.

I stole a bunch of code from rioxarray and now raster is a poor man's
version of rioxarray.  Except that eval() always returns a numpy array.