Initial version of rasterset with new Raster implementation
This commit adds a new method to Rasterset (creatively named build2 ATM)
that generates a chunked Dask Array for the output of the computation.
When the user calls load() on the resulting graph, the computation is
performed block-wise.  For each block all the input rasters get read and
then the respective functions evaluated.  Or, at least, that's the
theory.  I haven't yet tried to verify the block-wise operation.