Round bounds of output to multiples of resolution
After computing the intersection of all the source rasters and the shape
bounds (if any), round the bounds to be multiples of the resolution.  I
was doing this further down by rounding the window offsets and lengths
but that left the extent of the raster as unaligned.  With this change
the boundary will always be a multiple of the resolution.